new calendar using firebase as backend, events page refreshes automatically whenever there's a new event added
