+++ src/components/NavigationBar.vue
@@ -18,7 +18,7 @@
       <!-- Conditional links for authenticated users -->
       <router-link to="/Home" class="nav-item" v-if="isAuthenticated">Home</router-link>
       <router-link to="/Farm" class="nav-item" v-if="isAuthenticated">Farm</router-link>
-      <router-link to="/CalendarLanding" class="nav-item" v-if="isAuthenticated">Calendar</router-link>
+      <router-link to="/Events" class="nav-item" v-if="isAuthenticated">Calendar</router-link>
       <router-link to="/Settings" class="nav-item" v-if="isAuthenticated">Settings</router-link>
       <span class="vertical-line" v-if="isAuthenticated"></span>
